Pages/Optimization/Routing
Added:
Settings page,
User page,
Charts page.

Moved most of the stuff that was in App over to main instead to handle routing. First thought was to handle routing through the navbar itself but turns out if you do that you create a new router instance every time the navbar is rendered.. no bueno.

Anyways the App.tsx file is basically just serving as the home page of the app while main takes care of the heavy lifting. I could rename it to Home.tsx but I'm afraid of the red squiggly demons.